+++ jwst/extract_1d/extract.py
@@ -1011,7 +1011,7 @@ def create_poly(coeff):
         key = "c{}".format(i)
         coeff_dict[key] = coeff[i]
 
-    return polynomial.Polynomial1D(degree=n - 1, **coeff_dict)
+    return polynomial.Polynomial1D(degree=n-1, **coeff_dict)
 
 
 class ExtractBase:
@@ -1127,6 +1127,7 @@ class ExtractBase:
         self.independent_var = "pixel"
         self.src_coeff = None
         self.bkg_coeff = None
+        self.subtract_background = None
         self.p_src = None
         self.p_bkg = None
         self.smoothing_length = 0
@@ -1299,7 +1300,7 @@ class ExtractBase:
 class ExtractModel(ExtractBase):
     """The extraction region was specified in a JSON file."""
 
-    def __init__(self, input_model, slit,
+    def __init__(self, input_model, slit, verbose,
                  ref_file_type=None,
                  match="unknown",
                  dispaxis=HORIZONTAL, spectral_order=1,
@@ -1308,7 +1309,8 @@ class ExtractModel(ExtractBase):
                  independent_var="pixel",
                  smoothing_length=0, bkg_order=0, nod_correction=0.,
                  x_center=None, y_center=None,
-                 inner_bkg=None, outer_bkg=None, method='subpixel'):
+                 inner_bkg=None, outer_bkg=None, method='subpixel',
+                 subtract_background=None):
         """Create a polynomial model from coefficients.
 
         Extended summary
@@ -1400,6 +1402,12 @@ class ExtractModel(ExtractBase):
 
         method : str
             This is not relevant; it's only used for IFU data.
+
+        subtract_background : bool or None
+            A flag which indicates whether the background should be subtracted.
+            If None, the value in the extract_1d reference file will be used.
+            If not None, this parameter overrides the value in the
+            extract_1d reference file.
         """
 
         super().__init__()
@@ -1488,6 +1496,22 @@ class ExtractModel(ExtractBase):
         # corresponding polynomial functions.
         self.src_coeff = copy.deepcopy(src_coeff)
         self.bkg_coeff = copy.deepcopy(bkg_coeff)
+
+        if subtract_background is not None:
+            self.subtract_background = subtract_background
+            if subtract_background:
+                if self.bkg_coeff is None:
+                    self.subtract_background = False
+                    if verbose:
+                        log.info("Skipping background subtraction because "
+                                 "background regions are not defined.")
+            else:
+                if self.bkg_coeff is not None:
+                    self.bkg_coeff = None
+                    if verbose:
+                        log.info("Background subtraction will not be done; "
+                                 "it was specified in the reference file, but "
+                                 "it was overridden by the step parameter.")
 
         # These functions will be assigned by assign_polynomial_limits.
         # The "p" in the attribute name indicates a polynomial function.
@@ -1924,14 +1948,15 @@ class ExtractModel(ExtractBase):
 class ImageExtractModel(ExtractBase):
     """This uses an image that specifies the extraction region."""
 
-    def __init__(self, input_model, slit,
+    def __init__(self, input_model, slit, verbose,
                  ref_file_type=None,
                  match="unknown",
                  spectral_order=1,
                  ref_image=None,
                  dispaxis=HORIZONTAL,
                  smoothing_length=0,
-                 nod_correction=0):
+                 nod_correction=0,
+                 subtract_background=None):
         """Extract using a reference image to define the extraction and
            background regions.
 
@@ -2002,6 +2027,7 @@ class ImageExtractModel(ExtractBase):
                         smoothing_length)
             smoothing_length += 1               # must be odd
         self.smoothing_length = smoothing_length
+        self.subtract_background = subtract_background
 
         if self.exp_type == "NIS_SOSS":
             if hasattr(input_model.meta, 'wcs'):
@@ -2114,7 +2140,7 @@ class ImageExtractModel(ExtractBase):
             The background count rate that was subtracted from the total
             source count rate to get `net`.
 
-        npixels : ndarray, 1-D, float32
+        npixels : ndarray, 1-D, float64
             The number of pixels that were added together to get `net`.
 
         dq : ndarray, 1-D, uint32
@@ -2145,6 +2171,16 @@ class ImageExtractModel(ExtractBase):
         temp = np.ones_like(data)
         npixels = (temp * mask_target).sum(axis=axis, dtype=np.float)
 
+        if self.subtract_background is not None:
+            if not self.subtract_background:
+                if verbose and mask_bkg is not None:
+                        log.info("Background subtraction was turned off "
+                                 "- skipping it.")
+                mask_bkg = None
+            else:
+                if verbose and mask_bkg is None:
+                        log.info("Skipping background subtraction because "
+                                 "background regions are not defined.")
         # Extract the background.
         if mask_bkg is not None:
             n_bkg = mask_bkg.sum(axis=axis, dtype=np.float)
@@ -2414,10 +2450,10 @@ def interpolate_response(wavelength, relsens, verbose):
 
     Returns
     -------
-    r_factor : ndarray, 1-D
-        The response, interpolated at `wavelength`, with extrapolated
-        elements and zero or negative response values set to 1.  Divide
-        the net count rate by r_factor to obtain the flux.
+    rr_factor : ndarray, 1-D
+        The reciprocal of the response, interpolated at `wavelength`, with
+        extrapolated elements and zero or negative response values set to 0.
+        Multiply the net count rate by rr_factor to obtain the flux.
     """
 
     # "_relsens" indicates that the values were read from the RELSENS table.
@@ -2449,29 +2485,33 @@ def interpolate_response(wavelength, relsens, verbose):
     # `r_factor` is the response, interpolated at the wavelengths in the
     # science data.  -2048 is a flag value, to check for extrapolation.
     r_factor = np.interp(wavelength, wl_relsens, resp_relsens, -2048., -2048.)
-    mask = np.where(r_factor == -2048.)
-    if len(mask[0]) > 0:
+    mask2048 = np.where(r_factor == -2048.)
+    if len(mask2048[0]) > 0:
         if verbose:
-            log.warning("Using RELSENS, %d elements were extrapolated; "
-                        "these values will be set to 1.", len(mask[0]))
-        r_factor[mask] = 1.
-    mask = np.where(r_factor <= 0.)
-    if len(mask[0]) > 0:
+            log.warning("Using RELSENS, %d elements were extrapolated; the "
+                        "corresponding flux will be set to 0.",
+                        len(mask2048[0]))
+        r_factor[mask2048] = 1.                 # temporary
+    mask_neg = np.where(r_factor <= 0.)
+    if len(mask_neg[0]) > 0:
         if verbose:
             log.warning("Using RELSENS, %d interpolated response values "
-                        "were <= 0; these values will be set to 1.",
-                        len(mask[0]))
-        r_factor[mask] = 1.
+                        "were <= 0; the corresponding flux will be set to 0.",
+                        len(mask_neg[0]))
+        r_factor[mask_neg] = 1.                 # temporary
 
-    return r_factor
+    rr_factor = 1. / r_factor
+    rr_factor[mask2048] = 0.
+    rr_factor[mask_neg] = 0.
+
+    return rr_factor
 
 
-def do_extract1d(input_model, refname, smoothing_length, bkg_order,
-                 log_increment):
+def run_extract1d(input_model, refname, smoothing_length, bkg_order,
+                  log_increment, subtract_background):
     """Extract 1-D spectra.
 
-    Extended summary
-    ----------------
+    This just reads the reference file (if any) and calls do_extract1d.
 
     Parameters
     ----------
@@ -2493,11 +2533,110 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
         multi-integration, a message will be written to the log every
         `log_increment` integrations.
 
+    subtract_background : bool or None
+        User supplied flag indicating whether the background should be
+        subtracted.
+        If None, the value in the extract_1d reference file will be used.
+        If not None, this parameter overrides the value in the
+        extract_1d reference file.
+
     Returns
     -------
     output_model : data model
         A new MultiSpecModel containing the extracted spectra.
     """
+
+    # Read and interpret the reference file.
+    ref_dict = load_ref_file(refname)
+
+    # This item is a flag to let us know that do_extract1d was called
+    # from run_extract1d; that is, we don't expect this key to be present
+    # in ref_dict if do_extract1d was called directly.
+    # If this key is not in ref_dict, or if it is but it's True, then
+    # we'll set S_EXTR1D to 'COMPLETE'.
+    if ref_dict is not None:
+        ref_dict['need_to_set_to_complete'] = False
+
+    output_model = do_extract1d(input_model, ref_dict,
+                                smoothing_length, bkg_order,
+                                log_increment, subtract_background)
+
+    return output_model
+
+
+def ref_dict_sanity_check(ref_dict):
+    """Check for required entries.
+
+    Parameters
+    ----------
+    ref_dict : dict or None
+        The contents of the reference file.
+
+    Returns
+    -------
+    ref_dict : dict or None
+    """
+
+    if ref_dict is None:
+        return ref_dict
+
+    if 'ref_file_type' not in ref_dict:
+        # We can make an educated guess as to what this must be.
+        if 'ref_model' in ref_dict:
+            log.info("Assuming reference file type is image")
+            ref_dict['ref_file_type'] = FILE_TYPE_IMAGE
+        else:
+            log.info("Assuming reference file type is JSON")
+            ref_dict['ref_file_type'] = FILE_TYPE_JSON
+            if 'apertures' not in ref_dict:
+                raise RuntimeError("Key 'apertures' must be present in "
+                                   "the reference file.")
+            for aper in ref_dict['apertures']:
+                if 'id' not in aper:
+                    log.warning("Key 'id' not found in aperture {} "
+                                "in reference file".format(aper))
+
+    return ref_dict
+
+
+def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
+                 log_increment, subtract_background):
+    """Extract 1-D spectra.
+
+    Parameters
+    ----------
+    input_model : data model
+        The input science model.
+
+    ref_dict : dict or None
+        The contents of the reference file.  This will be None if there
+        is no reference file (i.e. if refname was "N/A").
+
+    smoothing_length : int
+        Width of a boxcar function for smoothing the background regions.
+
+    bkg_order : int
+        Polynomial order for fitting to each column (or row, if the
+        dispersion is vertical) of background.
+
+    log_increment : int
+        if `log_increment` is greater than 0 and the input data are
+        multi-integration, a message will be written to the log every
+        `log_increment` integrations.
+
+    subtract_background : bool or None
+        User supplied flag indicating whether the background should be subtracted.
+        If None, the value in the extract_1d reference file will be used.
+        If not None, this parameter overrides the value in the
+        extract_1d reference file.
+
+    Returns
+    -------
+    output_model : data model
+        A new MultiSpecModel containing the extracted spectra.
+    """
+
+    ref_dict = ref_dict_sanity_check(ref_dict)
 
     output_model = datamodels.MultiSpecModel()
     if hasattr(input_model, "int_times"):
@@ -2510,9 +2649,6 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
     # This will be relevant if we're asked to extract a spectrum and the
     # spectral order is zero.  That's only OK if the disperser is a prism.
     prism_mode = is_prism(input_model)
-
-    # Read and interpret the reference file.
-    ref_dict = load_ref_file(refname)
 
     if isinstance(input_model, datamodels.MultiSlitModel) or \
        isinstance(input_model, datamodels.MultiProductModel):
@@ -2537,6 +2673,8 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                                 ref_dict,
                                 slit, slit.name, sp_order,
                                 input_model.meta, smoothing_length, bkg_order)
+            if subtract_background is not None:
+                extract_params['subtract_background'] = subtract_background
             if extract_params['match'] == NO_MATCH:
                 log.critical('Missing extraction parameters.')
                 raise ValueError('Missing extraction parameters.')
@@ -2565,8 +2703,9 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
             if got_relsens and len(relsens) == 0:
                 got_relsens = False
             if got_relsens:
-                r_factor = interpolate_response(wavelength, relsens, True)
-                flux = net / r_factor
+                # reciprocal of the response
+                rr_factor = interpolate_response(wavelength, relsens, True)
+                flux = net * rr_factor
             else:
                 log.warning("No relsens for current slit, "
                             "so can't compute flux.")
@@ -2625,6 +2764,8 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                                     input_model, slitname, sp_order,
                                     input_model.meta, smoothing_length,
                                     bkg_order)
+                if subtract_background is not None:
+                    extract_params['subtract_background'] = subtract_background
                 if extract_params['match'] == EXACT:
                     slit = DUMMY
                     find_dispaxis(input_model, slit, sp_order, extract_params)
@@ -2655,8 +2796,9 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                 if got_relsens and len(relsens) == 0:
                     got_relsens = False
                 if got_relsens:
-                    r_factor = interpolate_response(wavelength, relsens, True)
-                    flux = net / r_factor
+                    # reciprocal of the response
+                    rr_factor = interpolate_response(wavelength, relsens, True)
+                    flux = net * rr_factor
                 else:
                     log.warning("No relsens for input file, "
                                 "so can't compute flux.")
@@ -2705,6 +2847,8 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                                     input_model, slitname, sp_order,
                                     input_model.meta, smoothing_length,
                                     bkg_order)
+                if subtract_background is not None:
+                    extract_params['subtract_background'] = subtract_background
                 if extract_params['match'] == NO_MATCH:
                     log.critical('Missing extraction parameters.')
                     raise ValueError('Missing extraction parameters.')
@@ -2747,10 +2891,11 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                         log.info(str(e) + ", skipping ...")
                         break
                     if got_relsens:
-                        r_factor = interpolate_response(
+                        # reciprocal of the response
+                        rr_factor = interpolate_response(
                                         wavelength, input_model.relsens,
                                         verbose)
-                        flux = net / r_factor
+                        flux = net * rr_factor
                     else:
                         flux = np.zeros_like(net)
                     fl_error = np.ones_like(net)
@@ -2805,7 +2950,8 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
                 source_type = input_model.meta.target.source_type.lower()
             except AttributeError:
                 source_type = "unknown"
-            output_model = ifu.ifu_extract1d(input_model, refname, source_type)
+            output_model = ifu.ifu_extract1d(input_model, ref_dict,
+                                             source_type, subtract_background)
 
         else:
             log.error("The input file is not supported for this step.")
@@ -2825,6 +2971,10 @@ def do_extract1d(input_model, refname, smoothing_length, bkg_order,
     # If the reference file is an image, explicitly close it.
     if ref_dict is not None and 'ref_model' in ref_dict:
         ref_dict['ref_model'].close()
+
+    if (ref_dict is None or 'need_to_set_to_complete' not in ref_dict or
+        ref_dict['need_to_set_to_complete']):
+            output_model.meta.cal_step.extract_1d = 'COMPLETE'
 
     return output_model
 
@@ -3181,12 +3331,12 @@ def extract_one_slit(input_model, slit, integ,
 
     if extract_params['ref_file_type'] == FILE_TYPE_IMAGE:
         # The reference file is an image.
-        extract_model = ImageExtractModel(input_model, slit, **extract_params)
+        extract_model = ImageExtractModel(input_model, slit, verbose, **extract_params)
         ap = None
     else:
         # If there is a reference file (there doesn't have to be), it's in
         # JSON format.
-        extract_model = ExtractModel(input_model, slit, **extract_params)
+        extract_model = ExtractModel(input_model, slit, verbose, **extract_params)
         ap = get_aperture(data.shape, extract_model.wcs,
                           verbose, extract_params)
         extract_model.update_extraction_limits(ap)
@@ -3272,7 +3422,7 @@ def nans_at_endpoints(wavelength, net, background, npixels, dq, verbose):
     background : ndarray
         Array of background values that were subtracted to get `net`.
 
-    npixels : ndarray, float32
+    npixels : ndarray, float64
         The number of pixels that were added together to get `net`.
 
     dq : ndarray