+++ jwst/extract_1d/ifu.py
@@ -1,4 +1,3 @@
-import json
 import logging
 import math
 
@@ -13,7 +12,7 @@ from . import spec_wcs
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
 
-def ifu_extract1d(input_model, refname, source_type):
+def ifu_extract1d(input_model, ref_dict, source_type, subtract_background):
     """Extract a 1-D spectrum from an IFU cube.
 
     Parameters
@@ -21,11 +20,17 @@ def ifu_extract1d(input_model, refname, source_type):
     input_model : JWST data model for an IFU cube (IFUCubeModel)
         The input model.
 
-    refname : string
-        The name of the JSON reference file.
+    ref_dict : dict
+        The contents of the reference file.
 
     source_type : string
         "point" or "extended"
+
+    subtract_background : bool or None
+        User supplied flag indicating whether the background should be subtracted.
+        If None, the value in the extract_1d reference file will be used.
+        If not None, this parameter overrides the value in the
+        extract_1d reference file.
 
     Returns
     -------
@@ -49,8 +54,9 @@ def ifu_extract1d(input_model, refname, source_type):
     if slitname is None:
         slitname = "ANY"
 
-    extract_params = ifu_extract_parameters(refname, slitname, source_type)
-
+    extract_params = get_extract_parameters(ref_dict, slitname, source_type)
+    if subtract_background is not None:
+        extract_params['subtract_background'] = subtract_background
     if extract_params:
         (ra, dec, wavelength, net, background, npixels, dq) = extract_ifu(
                         input_model, source_type, extract_params)
@@ -115,13 +121,13 @@ def ifu_extract1d(input_model, refname, source_type):
     return output_model
 
 
-def ifu_extract_parameters(refname, slitname, source_type):
+def get_extract_parameters(ref_dict, slitname, source_type):
     """Read extraction parameters for an IFU.
 
     Parameters
     ----------
-    refname : str
-        The name of the reference file.
+    ref_dict : dict
+        The contents of the reference file.
 
     slitname : str
         The name of the slit, or "ANY".
@@ -136,9 +142,7 @@ def ifu_extract_parameters(refname, slitname, source_type):
     """
 
     extract_params = {}
-    with open(refname) as f:
-        ref = json.load(f)
-    for aper in ref['apertures']:
+    for aper in ref_dict['apertures']:
         if 'id' in aper and aper['id'] != "dummy" and \
            (aper['id'] == slitname or aper['id'] == "ANY" or
             slitname == "ANY"):
@@ -232,6 +236,10 @@ def extract_ifu(input_model, source_type, extract_params):
 
     subtract_background = extract_params['subtract_background']
     smaller_axis = float(min(shape[1], shape[2]))       # for defaults
+    radius = None
+    inner_bkg = None
+    outer_bkg = None
+
     if source_type == 'point':
         radius = extract_params['radius']
         if radius is None:
@@ -259,10 +267,7 @@ def extract_ifu(input_model, source_type, extract_params):
         if height is None:
             height = smaller_axis / 2.
         theta = extract_params['theta'] * math.pi / 180.
-        radius = None
         subtract_background = False
-        inner_bkg = None
-        outer_bkg = None
 
     log.debug("IFU 1-D extraction parameters:")
     log.debug("  x_center = %s", str(x_center))