+++ scripts/runjwpsf.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 
-from __future__ import division, print_function  #confidence high
 import sys
 
 from jwst import jwpsf
+
 
 def run():
     jwpsf.jwpsf()